All customer endpoints added
There is currently no data validation or error handling in the code. A customer can be added, updated, queried and deleted using the endpoints.
